Vise utkast fra dokument, og trigge visning av utkast ved trykk på lik linje som i send brev. Flytter visning av lagrede utkast ut i egen komponent lagredeUtkast.tsx. Flyttet da også aktivtUtkast ut i redux slik at både sideDialogDokumenter.tsx og sendBrev.tsx kan trigge valgt utkast.
