+++ src/felleskomponenter/sideDialog/sendBrev/brevutkast/lagredeUtkast.less
@@ -1,0 +1,17 @@
+@import (reference) "~nav-frontend-core/less/_variabler";
+
+.lagredeUtkast {
+  background: @navLysGra;
+  border: 1px solid #d8d8d8;
+  padding: 1rem;
+  margin-bottom: 1rem;
+
+  &__utkast {
+    display: block;
+    margin: 0.125rem;
+
+    svg {
+      vertical-align: bottom;
+    }
+  }
+}
