+++ src/felleskomponenter/sideDialog/sendBrev/sendBrev.less
@@ -1,23 +1,5 @@
-@import (reference) "~nav-frontend-core/less/_variabler";
-
 .send_brev {
   font-family: "Source Sans Pro", Arial, sans-serif;
-
-  .brevutkast {
-    background: @navLysGra;
-    border: 1px solid #d8d8d8;
-    padding: 1rem;
-    margin-bottom: 1rem;
-
-    &__utkast {
-      display: block;
-      margin: 0.125rem;
-
-      svg {
-        vertical-align: bottom;
-      }
-    }
-  }
 
   .send_brev__apne-nytt-vindu-container {
     display: flex;
